% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sonde_data.R
\name{sonde_file_convert}
\alias{sonde_file_convert}
\title{Sonde data convert}
\usage{
sonde_file_convert(filename, folder, station, latitude, longitude)
}
\arguments{
\item{filename}{The name of the file, saved as an xlsx file. Don't include the file extension. Use quotation marks.}

\item{folder}{sub-folder where this file is saved. Sub-folder must be within "Documents." Use quotation marks. For ex: "/NT_Hydrology/Quality/"}

\item{station}{station abbreviation. Use quotation marks. For ex: "AKL-PCH"}

\item{latitude}{latitude of the station. Input number only, no quotation marks.}

\item{longitude}{longitude of the station. Input number only, no quotation marks.}
}
\description{
Function that renames column headings, manipulates date-time fields and re-formats column order
}
